Provide initial Getting Started Template
This is meant to be a minimalistic Asciidoctor project
to be cloned and used for Gradle Guides.

It has a Gradle build that compiles only the README.adoc file
and uses the gh-pages Gradle plugin for publishing to GitHub
Pages.

The README itself is a simple guide for creating one's own
guide for contributors.

See: gradle/build-tool-web#71