Push table of contents config into build script
This configuration is orthogonal to the content of the guide itself and
therefore better expressed at the level of the build. Other similar
configuration is already located there.

In general, we should favor an approach that leaves "content in the
docs and configuration in the build" unless we find compelling reasons
to do otherwise.

See gradle/build-tool-web#71