Remove "draft" sidebar
There is no need for an explicit "draft" status for these and other
guides. A guide is either linked to from the guides section of gradle.org
or it is not. Once it is linked, it is "published" and intended for
consumption by Gradle users. Beyond that, guides are in a perpetual
state of being subject to change as new versions of gradle distributions
and plugins get released, and as new best practices emerge. From this
perspective, there is little value in distinguishing between "draft" and
"final", as there is no "final" state per se. What matters is whether we
think the content is "good enough" to go live, and that is a function of
the review / editing process for each guide as it gets written.

See gradle/build-tool-web#71