Add sections and normalize section heading grammar
These changes ensure that we keep the section hierarchy as flat,
consistent and inviting as possible. For example, do not distinguish
between a "Prerequisites" section and an "Instructions" (or "Steps")
section with sub-sections under each.

The intention here is to keep these short and simple Getting Started
Guides *looking* short and simple to the reader. Formal "Introduction"
or "Prerequisite" sections and nested sections within risk communicating
complexity to the reader. Instead of a an explicit introduction /
prerequisite sections, favor simple, clear and concrete headings such
as:

 - What you'll build
 - What you'll need

Likewise, instead of a concrete "instructions" or "steps" headings, favor
headings written in the imperative that clearly instruct the user to do
something, such as:

 - Create a new repository
 - Generate and view your guide
 - Publish your guide to GitHub Pages

See gradle/build-tool-web#71