Complete the set of prerequisites
Always tell the user roughly how much time they should expect to spend
on a guide, e.g. "About 15 minutes". This communicates to the user that
we value their time and that we're not going to take much of it. As the
reader works through several guides and notices that our time estimates
are more or less right, this builds trust.

Don't assume anything about prerequisites. Tell the user they'll need
Git if they'll need Git, and mention that they'll need a text editor if
that's the case.

Add links wherever possible to instructions for downloading and
installing or otherwise satisfying a given prerequisite. Remember, we're
addressing a very diverse audience. Not everyone knows how to install
Git or even what a JRE is.

By listing these prerequisites explicitly and exhaustively, we create
confidence in the user that the list is really accurate. If the reader
notices that we took the time to tell them that they need "A text
editor" and "a JRE"--items that many guides might leave out--they will
be inclined to believe that the list is complete, and that they won't
get surprised halfway through by getting told to use something they
don't have.

Listing the prerequisites explicitly in this fashion also keeps us
honest about whether the guide is actually simple. If the list of
prerequisites is 10 items long, something is probably wrong, e.g. the
guide may need to be split up, or more automation should be introduced
to provision some of the prerequisities on behalf of the user.

See gradle/build-tool-web#71