Favor permanent links to reference documentation
The 'current' symlink typically used in links to Gradle reference
documentation is fragile, because it can easily result in broken links
when the underlying reference documentation changes. We should favor
permalinks instead, so that we can "guarantee" links will continue to
work even across new Gradle releases. We still may need to change these
links, but this way it happens as part of a controlled and intentional
upgrade process for each guide. Not as a "surprise" event triggered by
the publication of a new Gradle release.

This commit also replaces explicit references to Gradle version numbers
with the {gradle-version} attribute for consistency.

See gradle/build-tool-web#12
See gradle/build-tool-web#92