Remove IDE tips until instructions are available
Avoid mentioning the IDE at all until we have a proper guide or other
resource that actually instructs the user what to do with their IDE to
succeed with these guides. Mentioning it without any further support
risks frustrating the reader, and gives us as authors a false sense of
confidence that we've somehow "addressed" this issue, when we actually
have not. We will properly address it in gradle/build-tool-web#90.

See gradle/build-tool-web#12