Link to user manual in 'Next Steps' where guides are not available
As a best practice (gradle/build-tool-web#92), make sure that the "Next
Steps" section is 100% usable by readers now, even if the resources we
link to are not ideal. In this case, we call out three possible "Next
Steps", each of which would ideally have their own Getting Started
Guides, and for each, we create a new card on the build-tool-web board
to track the creation of that guide:

 - Consuming JVM Libraries (gradle/build-tool-web#107)
 - Publishing JVM Libraries (gradle/build-tool-web#108)
 - Multi-Project Builds (gradle/build-tool-web#106)

In each of the above cases, we link to the best resource currently
available. When each guide comes online, we can return to this guide and
swap the links out for each respectively.

Additionally, we've identified the need for a "Getting Started Working
with Composite Builds" guide, and created a card for it at
gradle/build-tool-web#105, but have *not* added a bullet point for it in
this guide, because there is no generally available Gradle distribution
that has support for composite builds as yet. When that happens (in
Gradle 3.1), we can consider adding a bullet point for it, and a link to
the best available section of the user manual (if we have not already
written the new guide by that time).

See gradle/build-tool-web#12