Add an exclamation point to the Greeter's greeting
Because this is exciting! Seriously, though, it better follows
convention with the classic "Hello, World!", and as the final
instruction and output in the guide, the exclamation point ends the user
experience on a high note. Too many exclamation points are a bad thing,
a single one used well can have a nice effect.

See gradle/build-tool-web#12
See gradle/build-tool-web#92