Polish prose and source listings
These changes update the prose of the text for:

 - Consistency with changes to headings made in the previous commit
 - Grammar, especially where giving instructions to the reader is
   concerned

A source listing has been added in favor of a mixing `literal`
commands within prose bullet points.

The `clean` task has been eliminated from the "Customize the library
JAR" section, as the instructions and results still work perfectly well
without it. Yes, it results in two JAR files residing under build/libs,
but this fact is minor and not something that the user will stumble upon
or struggle with if following the instructions as given. Since
explaining the use of `clean` would be a digression of marginal value,
and since the absence of `clean` does not create any real problem in the
experience of completing the guide, this commit errs on the side of
eliminating it.

In general, these changes focus on delivering a consistent, pleasant
reading experience and a clear, unambiguous instruction-following
experience.

See gradle/build-tool-web#12
See gradle/build-tool-web#92