Complete the writing a plugin GS guide
This fleshes out the prose from the previous draft, removes the use of an
extension from the sample plugin, and reintroduces a plugin identifier.

Once gradle/guides#30 is done, this guide should ideally link to the Writing
a Custom Task GS guide instead of the user manual.

Issue: gradle/guides#9