Update guide for new submission process
We have changed the process for working with external contributors on Getting
Started Guides, which is reflected in these modifications. The author should
now raise an issue describing the proposal, and once accepted, they will fork
a repository that we create for them, write the guide, and then submit it via
a pull request.

Issue: gradle/guides#8