Create (very rough) draft outline (gradle/guides#30)
This is really a rough outline. Most of the information is hidden
inside asciidoc comment blocks. The draft follows the journey of
creating a basic custom tasks then enhancing it with inputs and
outputs.

As we still had some discussion of where we wanted to go with this
this commit leaves it more at a conversation starter point than a
full outline.

Many parts that are in comment blocks are hard-wrapped as they will
be deleted when the content is plugged in.

Issue: gradle/guides#30