First draft of 'Building Groovy libraries' (gradle/guides#76)
Based upon similar structure to that of 'Building Java Applications'.

This extends the concept by using automation to run the init task as part of
the build process and then annotating the source code on the fly before inclusion
into the Asciidoc. This will allow for ease of maintenance when upgrading the
Gradle version in future for this guide.