Move daemon chapter down
The daemon is on by default since 3.0. For users on older versions the
recommendation is still valid, but everyone else can safely ignore it.

Also removed the warning about calling out to other processes. This is
a plugin development issue and a much lower-level topic than should be
mentioned in an "easy wins" section.