Add script calculating time to complete of a guide
This script is used in the <script> element of every guide page. Its
purpose is to:

 - Compute an approximate value in minutes of the time required to
   read an entire guide (e.g. "15").

 - Find <span> elements with the class "time-to-complete-text" and add
   time to complete text to those <span> elements (e.g. "15 minutes").

To use the script, add the following code to a guide's AsciiDoc file:

 +++<span class="time-to-complete"></span>+++

The `+++` symbols are required to prevent Asciidoctor from substituting
characters like `<`. For more information on this syntax, see
"Triple-plus passthrough" at asciidoctor.org/docs/asciidoc-writers-guide

Issue: #82