Remove the services from the jar files
To get this to run with the gradle run task or the generated start
script, we also have add the `--module-path` and `-m` options to the
jvm arguments.

Since we cannot add these arguments to the tail end of the
`jvmArgs` and `defaultJvmOpts` lists, we have to repeat the
`mainClassname` in the `jvmArgs` and `defaultJvmOpts` lists.