Update the gradle wrapper version
We need to use the nightly version of Gradle so that the instructions
will say that that's the version of gradle needed to run through this
guide.

By the time the guide is released, we should be on Gradle 4.1, so the
wrapper script can be updated again to use that.