Fix windows build for building shared libraries (gradle/guides#136)
In Windows, when using Microsoft VC++, the method declaration of library
methods should have __declspec(dllexport) to define the storage class.
This is specific to MS VC++ and when using that compiler, the method
declations should be modified accordingly. This commit updates the
build logic to define DLL_EXPORT if it is VC++ tool chain and the
greeter.hpp defines a DECLSPEC macro which is set only if DLL_EXPORT
is defined. Otherwise, it stays empty.