Getting started with testing c++ libraries (gradle/guides#136)
This getting started guide now shows how build and test c++ libraries.
This uses GoogleTest and the google-test plugin to building and running
the tests againt the c++ library that is being built. The build.gradle has
been updated to fetch and build GoogleTest library for windows/linux and
the cpp.gradle has been updated to declare the GoogleTest as a pre-built
library and statically link the test suites against the GoogleTest library.