Added a set of Spock tests (gradle/guides#149)
Validate all steps and corresponding output with a Spock test fixture.

I have placed one gradleRunner test in place purely as comparison as to
how the same kind of otput validation can be achieved with GraddleRunner.
This can be removed at some stage, but it think it can serve as a
discussion point for the immediate future.