Replace index.html with nginx autoindexing
We will likely forget to update index.html whenever adding, removing or
renaming a guide or guide URL. For this reason, let's use nginx
autoindexing to list all the files and directories in this repository at
http://localhost:8080. Previewing the "Building Android Apps" guide is
as simple as clicking on the "building-android-apps/" directory. There's
no need to maintain a list of guides like we planned to do in
index.html.