Remove mention of CoD
Almost no user ever gets any benefit from configure-on-demand,
because it requires you to structure your project completely diffently
from the widespread conventions. Just activating it on your average
Gradle project will have no benefit, but can lead to lots of quirks with
plugins that are incompatible with that option.