Improve display of multi-file, multi-language samples
This change changes the titles of language selectors to the languages
themselves, instead of using the file names.

In addition, switching all of the samples in a document no longer
scrolls the page. This is achieved by recording the content offset
height before and after the language switch, and scrolling the window
via JS.

Issues:
gradle-guides/gradle-guides-plugin#28
gradle-guides/gradle-guides-plugin#29