Make it clear that reader must start with a plugin project
For readers coming from the Writing Gradle Plugins guide, it wasn't clear that
the project they had been working on wasn't immediately suitable for this guide
because the plugin code is located in the `buildSrc` directory. This change
should make it much clearer as to the starting point for the guide. In
particular, it now actively encourages the reader to use the example plugin
project that we have on GitHub.

Attempts to resolve https://github.com/gradle/guides/issues/187.

Signed-off-by: Peter Ledbrook <peter@gradle.com>