Don't cache too much on Travis
Travis is cloning the repository every time so the creation time changes, it make fileHashes to change, and it creates a chain of changes..
Before Gradle 4.10 it worked fine but as I told @oehme now it's not working anymore.

On Travis even if a single file changed the cache will be tar and uploaded, so as the size of the cache grows it can take longer than compile.

Signed-off-by: Cristian Garcia <cristiangm@users.noreply.github.com>