Migrate legacy sample to documentation sample plugin
This PR grew to be bigger than expected. It migrate some of the legacy
logic to the new documentation plugin. It also removed some
configuration that doesn't have any tests and is not used. Basically
those are modeling that went too far. Some modeling were also moved to
the internal API when it's not used anywhere.